requirements.txt
#pattern recognition
Ta-Lib==0.4.28
numpy==1.26.4
pandas==2.2.1
requests==2.31.0

#youtube
google-api-python-client==2.123.0
youtube-transcript-api==0.6.2
#deepgram-sdk==3.2.2
cohere==5.1.2
openai==1.14.3
newsapi-python==0.2.7

#global
fastapi==0.110.0
uvicorn[standard]

